+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # Changelog
 
+## 1.3.0 - 2026-08-03
+
+### Added
+
+- Add dark param to gg console colors
+
 ## 1.2.0 - 2026-08-03
 
 ### Added
 
 - Add .gitattributes file
+- Add a `dark` param printing the message in `darkGray`. It is ignored when
+`colorize` is false
 
 ### Changed